overall cleaning and reformating
converter.jar was recompiled with the new changes.

Ass.java was cleaned and reformatted.

Dfxp.java was modified to write the sub file in the right encoding

Entery.java was changed from filereader library to bufferdreader and you can choose what encoding the sub file is using .

Srt.java was cleaned and reformatted.